requirements.txt
apache-airflow[crypto]
psycopg2-binary
werkzeug
flask-bcrypt